Revert "Rocket booking cancelations actions"
